backend/requirements.txt
djangorestframework
pillow
djangorestframework-simplejwt
django
django-cors-headers